+++ packages/koa/src/jsonapi.spec.ts
@@ -6,16 +6,6 @@ import Koa from "koa"
 import { Hatchify } from "./koa"
 import { GET, POST, createServer } from "./testing/utils"
 
-const ERROR_CODE_MISSING_DATA = {
-  status: 422,
-  code: "missing-data",
-  title: " 'data' must be specified for this operation. ",
-  detail: "payload was missing 'data' field. It can not be null/undefined.",
-  source: {
-    pointer: "/data",
-  },
-}
-
 describe("JSON:API Tests", () => {
   const Model: HatchifyModel = {
     name: "Model",
@@ -28,15 +18,6 @@ describe("JSON:API Tests", () => {
         type: DataTypes.STRING,
         allowNull: false,
       },
-    },
-  }
-
-  const TodoModel = {
-    name: "Todo",
-    attributes: {
-      name: "STRING",
-      due_date: "DATE",
-      importance: "INTEGER",
     },
   }
 
@@ -100,42 +81,5 @@ describe("JSON:API Tests", () => {
     expect(find.deserialized.id).toBe(create.deserialized.id)
 
     await hatchify.orm.close()
-  })
-
-  it("should return error 422 when invalid data schema is passed", async () => {
-    const app = new Koa()
-    const hatchify = new Hatchify([TodoModel], {
-      prefix: "/api",
-      database: {
-        dialect: "sqlite",
-        storage: "example.sqlite",
-      },
-    })
-
-    app.use(hatchify.middleware.allModels.all)
-    const server = createServer(app)
-    await hatchify.createDatabase()
-
-    const create = await POST(
-      server,
-      "/api/todos",
-      {
-        invalid: {
-          type: "Todo",
-          attributes: {
-            id: "101",
-            name: "Walk the dog",
-            due_date: "2024-12-12",
-            importance: 6,
-          },
-        },
-      },
-
-      "application/vnd.api+json",
-    )
-
-    expect(create).toBeTruthy()
-    expect(create.status).toBe(422)
-    expect(create.deserialized).toContainEqual(ERROR_CODE_MISSING_DATA)
   })
 })
